+++ tests/gui/utils/test_macros.h
@@ -43,15 +43,6 @@
 
 #define ADD_SEMI_LOG_Y ADD_PLOT_TYPE(cmp::SemiLogY)
 
-#define PLOT_Y(Y) \
-  { GET_PLOT->plot(Y); }
-
-#define PLOT_XY(Y, X) \
-  { GET_PLOT->plot(Y, X); }
-
-#define PLOT_XY_ATTRI(Y, X, ATTRI) \
-  { GET_PLOT->plot(Y, X, ATTRI); }
-
 #define REALTIMEPLOT(Y) \
   { GET_PLOT->realTimePlot(Y); }